Fix all warnings reported in issues (and then some)

+++ igloo/core/testresults.h
@@ -20,12 +20,12 @@ namespace igloo {
 
       int NumberOfSucceededTests() const
       {
-        return succeededTests_.size();
+        return static_cast<int>(succeededTests_.size());
       }
 
       int NumberOfFailedTests() const
       {
-        return failedTests_.size();
+        return static_cast<int>(failedTests_.size());
       }
 
       void AddResult(const SucceededTestResult result)
